# lib/api-zod/src/generated/types/voterProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Common Ground — match voters with candidates by values
 * OpenAPI spec version: 0.1.0
 */
import type { IssueStance } from './issueStance';
import type { VoterLocation } from './voterLocation';

export interface VoterProfile {
  hasProfile: boolean;
  completedOnboarding: boolean;
  location?: VoterLocation | null;
  stances: IssueStance[];
  answeredCount: number;
  totalQuestions: number;
  /** @nullable */
  updatedAt?: string | null;
}